# docs/openapi/quality-control.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quality Control API",
    "version": "2026-07-07",
    "description": "Documentacao inicial dos endpoints criticos do Quality Control. Esta especificacao deve ser atualizada no mesmo PR em que um endpoint novo for criado ou uma regra de permissao for alterada."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesmo host da aplicacao Next.js"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Sessao web autenticada usada pelas rotas protegidas."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "additionalProperties": true
      },
      "AdminUserSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "user": { "type": ["string", "null"] },
          "role": { "type": ["string", "null"] },
          "companyId": { "type": ["string", "null"] },
          "companySlug": { "type": ["string", "null"] }
        },
        "additionalProperties": true
      },
...
    }
  },
  "security": [
    {
      "sessionAuth": []
    }
  ],
  "paths": {
    "/api/admin/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Lista usuarios visiveis no escopo atual",
        "description": "Leader TC e perfis globais veem escopo amplo. Empresa ve apenas usuarios da propria empresa. Sem permissao o endpoint retorna 403.",
        "x-authorizedRoles": ["leader_tc", "technical_support", "empresa"],
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filtra a listagem pela empresa desejada quando o escopo do usuario permitir."
          },
          {
            "name": "summary",
            "in": "query",
            "schema": { "type": "string", "enum": ["login", "logins", "identity"] },
            "description": "Retorna apenas resumo de identidade/login."
          }
        ],
        "responses": {
          "200": {
            "description": "Usuarios listados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AdminUserSummary" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Nao autenticado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Sem permissao ou fora do escopo de empresa", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Cria usuario no contexto operacional atual",
        "description": "Empresa pode criar apenas usuario empresarial dentro da propria empresa. Perfis globais podem criar conforme regra de perfil.",
        "x-authorizedRoles": ["leader_tc", "technical_support", "empresa"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "role": { "type": "string" },
                  "client_id": { "type": ["string", "null"] },
                  "clientSlug": { "type": ["string", "null"] },
                  "companySlug": { "type": ["string", "null"] },
                  "capabilities": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                }
              },
              "example": {
                "name": "Usuaria Empresa QA",
                "email": "qa.empresa@example.com",
                "role": "company_user",
                "clientSlug": "testing-company"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario criado com sucesso" },
          "400": { "description": "Payload invalido" },
          "401": { "description": "Nao autenticado" },
          "403": { "description": "Sem permissao para criar o perfil solicitado" },
          "409": { "description": "Conflito de empresa ou usuario" }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Atualiza dados do usuario a partir da tela administrativa",
        "description": "Mantem as mesmas regras de escopo e perfil da listagem/criacao.",
        "x-authorizedRoles": ["leader_tc", "technical_support", "empresa"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "role": { "type": "string" },
                  "active": { "type": "boolean" },
                  "capabilities": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                },
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Usuario atualizado" },
          "401": { "description": "Nao autenticado" },
          "403": { "description": "Sem permissao" },
          "404": { "description": "Usuario nao encontrado" }
        }
      }
    },
...
  }
}